finished refractoring, now need to put in the graphql mutations
