+++ frontend/src/components/ChangeMeetingStatusModal/ChangeMeetingStatusOuterModal/ChangeMeetingStatusOuterModal.scss
@@ -1,0 +1,27 @@
+@import '../../../utils/colors';
+
+.ChangeMeetingStatusOuterModal{
+    .modalAgendaItemWrapper{
+        .AgendaItem{
+            border:solid 4px $darkBlue;
+        }
+    }
+    
+    .closeOutModal{
+        position: absolute;
+        color: white;
+        fill: white;
+        width:100%;
+        cursor:pointer;
+
+        .closeOutText{
+            position: relative;
+            bottom: 0.4rem;
+            font-weight: 600;
+        }
+    }
+    
+
+    @import '../../../utils/statusButtons.scss';
+
+}
